spliting reducer
