+++ prepare/front/components/LoginForm.js
@@ -2,11 +2,10 @@ import React, { useCallback } from "react";
 import { Button, Form, Input } from "antd";
 import Link from "next/link";
 import styled from "styled-components"; // 태그 안에 객체 생성 뒤 style을 설정하면 렌더링 시 새로운 객체로 인식한다 때문에 styled-components 사용
-import PropTypes from "prop-types";
 //* customHook 적용하기
 import useInput from "../hooks/useInput"; // 패턴이 비슷한데 조금씩 다른 것들 커스텀 훅으로 처리할 수 있다
 import { useDispatch } from "react-redux"; // useDispatch()
-import { loginAction } from "../reducers";
+import { loginAction } from "../reducers/user";
 
 const LoginForm = () => {
   const dispatch = useDispatch();
