+++ prepare/front/reducers/index.js
@@ -1,63 +1,28 @@
-import { HYDRATE } from "next-redux-wrapper";
-
-// reducer
-const initialState = {
-  user: {
-    isLoggedIn: false,
-    user: null,
-    signUpData: {},
-    loginData: {},
-  },
-  post: {
-    main: [],
-  },
-};
+import { HYDRATE } from "next-redux-wrapper"; // SSR(Server Side Rendering)을 위함
+import { combineReducers } from "redux"; // combineReducers은 reducer 함수를 합쳐주는 역할
+// spliting reducer
+import user from "./user";
+import post from "./post";
 
 // async action creator(비동기) / redux-saga
 
-// action creactor
-export const loginAction = (data) => {
-  return {
-    type: "LOG_IN",
-    data,
-  };
-};
-export const logoutAction = () => {
-  return {
-    type: "LOG_OUT",
-  };
-};
-
 // (이전상태, 액션) >> 다음 상태
-const rootReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case HYDRATE:
-      console.log("HYDRATE", action);
-      return {
-        ...state,
-        ...action.payload,
-      };
-    case "LOG_IN":
-      return {
-        ...state,
-        user: {
-          ...state.user, // 바꾸지 않을 데이터는 참조관계 유지
-          isLoggedIn: true,
-          user: action.data,
-        },
-      };
-    case "LOG_OUT":
-      return {
-        ...state,
-        user: {
-          ...state.user,
-          isLoggedIn: false,
-          user: null,
-        },
-      };
-    default:
-      return state;
-  }
-};
+const rootReducer = combineReducers({
+  // HYDRATE를 위해 index reducer를 추가 (SSR을 위해 추가함)
+  index: (state = {}, action) => {
+    switch (action.type) {
+      case HYDRATE:
+        console.log("HYDRATE", action);
+        return {
+          ...state,
+          ...action.payload,
+        };
+      default:
+        return state;
+    }
+  },
+  user,
+  post,
+});
 
 export default rootReducer;
