+++ prepare/front/reducers/post.js
@@ -1,0 +1,12 @@
+export const initialState = {
+  mainPosts: [],
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
+
+export default reducer;
